revert: remove skill:command_registered hook listener from tool-slash-command
- Delete on_skill_command_registered handler function from tool.py
- Remove coordinator.hooks.register call for skill:command_registered event
- Remove unused CommandMetadata and ParsedCommand imports
- Delete obsolete test_skill_command_hook.py tests
- Add test_tool_mount_no_hooks.py verifying no hooks are registered after mount

Closes task-4-revert-slash-command-listener
